+++ src/diary/ops/mod.rs
@@ -3,3 +3,4 @@ pub use self::init::init;
 pub use self::init::InitOptions;
 pub mod add;
 pub mod new;
+pub mod open;