+++ src/diary/ops/add.rs
@@ -5,7 +5,6 @@
 use std::{fs::File, io, io::Write};
 
 use chrono::{DateTime, Local};
-use edit;
 
 use crate::{
     errors::DiaryError,
@@ -30,7 +29,7 @@ pub struct AddOptions<'a> {
 /// # Errors
 ///
 /// * If `file_result` contains an error.
-/// * If
+/// * If the content provided is empty.
 fn add_content(
     file_result: io::Result<File>,
     content: String,
@@ -67,63 +66,90 @@ fn add_content(
 /// # Returns
 ///
 /// The unit if adding the content was successful, a DiaryError otherwise.
-pub fn add(opts: &AddOptions, config: &Config, date: &DateTime<Local>) -> Result<(), DiaryError> {
+pub fn add(
+    opts: &AddOptions,
+    config: &Config,
+    date: &DateTime<Local>,
+    string_getter: editing::StringGetter,
+) -> Result<(), DiaryError> {
     let file_result =
         file_system::get_entry(config.diary_path().to_path_buf(), date, config.prefix());
 
-    let content = edit::edit("")?;
+    let content = string_getter("".to_string())?;
 
     add_content(file_result, content, opts.tag)
 }
 
 #[cfg(test)]
 mod test {
-    use std::io::Read;
-    use tempfile;
+    use chrono::{Local, TimeZone};
+    use std::fs;
+    use tempfile::tempdir;
 
-    use super::add_content;
-
+    use crate::{
+        ops::{
+            add::{add, AddOptions},
+            new::{new, NewOptions},
+        },
+        utils::editing::{test_empty_string_getter, test_string_getter},
+        Config,
+    };
     #[test]
-    fn add_content_no_tag() {
-        let tag = None;
-        let content = "Test content".to_string();
+    fn add_no_tag() {
+        let entry_date = Local.ymd(2021, 11, 6).and_hms(0, 0, 0);
+        let opts = AddOptions { tag: None };
+        let temp_dir = tempdir().unwrap();
+        let mut filepath = temp_dir.path().to_path_buf();
 
-        let mut file = tempfile::NamedTempFile::new().unwrap();
+        let config = Config::new(filepath.clone(), "diary".to_string());
 
-        let temp_file = file.reopen();
+        let new_opts = NewOptions { open: false };
 
-        add_content(temp_file, content, tag).unwrap();
+        new(&new_opts, &config, &entry_date, test_string_getter).unwrap();
 
-        let mut buffer = String::new();
-        file.read_to_string(&mut buffer).unwrap();
+        add(&opts, &config, &entry_date, test_string_getter).unwrap();
 
-        assert_eq!(buffer, "Test content\n");
+        filepath.push("2021-11/diary_2021-11-06.md");
+        let content = fs::read_to_string(filepath).unwrap();
+
+        assert!(content.contains("Test content"));
     }
 
     #[test]
-    fn add_content_with_tag() {
-        let tag = Some("Test");
-        let content = "Test content".to_string();
+    fn add_with_tag() {
+        let entry_date = Local.ymd(2021, 11, 6).and_hms(0, 0, 0);
+        let opts = AddOptions { tag: Some("Tag") };
+        let temp_dir = tempdir().unwrap();
+        let mut filepath = temp_dir.path().to_path_buf();
 
-        let mut file = tempfile::NamedTempFile::new().unwrap();
+        let config = Config::new(filepath.clone(), "diary".to_string());
 
-        let temp_file = file.reopen();
+        let new_opts = NewOptions { open: false };
 
-        add_content(temp_file, content, tag).unwrap();
+        new(&new_opts, &config, &entry_date, test_string_getter).unwrap();
 
-        let mut buffer = String::new();
-        file.read_to_string(&mut buffer).unwrap();
+        add(&opts, &config, &entry_date, test_string_getter).unwrap();
 
-        assert_eq!(buffer, "## Test\n\nTest content\n");
+        filepath.push("2021-11/diary_2021-11-06.md");
+        let content = fs::read_to_string(filepath).unwrap();
+
+        assert!(content.contains("Test content"));
+        assert!(content.contains("Tag"));
     }
     #[test]
     #[should_panic(expected = "value: NoContent")]
-    fn add_content_empty_string() {
-        let tag = Some("Test");
-        let content = "".to_string();
+    fn add_empty_string() {
+        let entry_date = Local.ymd(2021, 11, 6).and_hms(0, 0, 0);
+        let opts = AddOptions { tag: Some("Tag") };
+        let temp_dir = tempdir().unwrap();
+        let filepath = temp_dir.path().to_path_buf();
 
-        let file = tempfile::tempfile();
+        let config = Config::new(filepath, "diary".to_string());
 
-        add_content(file, content, tag).unwrap();
+        let new_opts = NewOptions { open: false };
+
+        new(&new_opts, &config, &entry_date, test_string_getter).unwrap();
+
+        add(&opts, &config, &entry_date, test_empty_string_getter).unwrap();
     }
 }