+++ src/diary/ops/new.rs
@@ -29,7 +29,12 @@ pub struct NewOptions {
 /// The unit upon successful creation of the entry.
 /// DiaryError if the entry already exists.
 /// DiaryError on any other IO issues.
-pub fn new(opts: &NewOptions, config: &Config, date: &DateTime<Local>) -> Result<(), DiaryError> {
+pub fn new(
+    opts: &NewOptions,
+    config: &Config,
+    date: &DateTime<Local>,
+    string_getter: editing::StringGetter,
+) -> Result<(), DiaryError> {
     config.initialised()?;
     let mut new_entry_path = file_system::month_folder(config.diary_path().to_path_buf(), date);
     file_system::create_month_folder(&new_entry_path)?;
@@ -50,7 +55,7 @@ pub fn new(opts: &NewOptions, config: &Config, date: &DateTime<Local>) -> Result
         Err(e) => return Err(e.into()),
     };
     if opts.open {
-        let contents = edit::edit("")?;
+        let contents = string_getter("".to_string())?;
         editing::add_user_content_to_file(&mut file, contents)?;
     };
     Ok(())
@@ -58,9 +63,9 @@ pub fn new(opts: &NewOptions, config: &Config, date: &DateTime<Local>) -> Result
 
 #[cfg(test)]
 mod test {
-    use std::path::PathBuf;
+    use std::{fs, path::PathBuf};
 
-    use crate::{ops::init, Config};
+    use crate::{ops::init, utils::editing::test_string_getter, Config};
     use chrono::prelude::*;
     use init::InitOptions;
     use tempfile::tempdir;
@@ -79,7 +84,7 @@ mod test {
 
         init(&init_opts, &config).unwrap();
 
-        new(&new_opts, &config, &date).unwrap();
+        new(&new_opts, &config, &date, test_string_getter).unwrap();
 
         let mut test_path = config.diary_path().clone();
         test_path.push("2021-11");
@@ -99,7 +104,7 @@ mod test {
         let config = Config::new(diary_path, String::from("diary"));
         let date = Local.ymd(2021, 11, 6).and_hms(0, 0, 0);
 
-        new(&new_opts, &config, &date).unwrap();
+        new(&new_opts, &config, &date, test_string_getter).unwrap();
     }
 
     #[test]
@@ -115,8 +120,8 @@ mod test {
 
         init(&init_opts, &config).unwrap();
 
-        new(&new_opts, &config, &date).unwrap();
-        new(&new_opts, &config, &date).unwrap();
+        new(&new_opts, &config, &date, test_string_getter).unwrap();
+        new(&new_opts, &config, &date, test_string_getter).unwrap();
     }
 
     #[test]
@@ -126,6 +131,27 @@ mod test {
         let config = Config::default();
         let date = Local.ymd(2021, 11, 6).and_hms(0, 0, 0);
 
-        new(&new_opts, &config, &date).unwrap();
+        new(&new_opts, &config, &date, test_string_getter).unwrap();
+    }
+
+    #[test]
+    fn new_open_file_success() {
+        let new_opts = NewOptions { open: true };
+        let init_opts = InitOptions {
+            path: PathBuf::from(""),
+        };
+        let diary_path = tempdir().unwrap().path().to_path_buf();
+        let config = Config::new(diary_path, String::from("diary"));
+        let date = Local.ymd(2021, 11, 6).and_hms(0, 0, 0);
+
+        init(&init_opts, &config).unwrap();
+
+        new(&new_opts, &config, &date, test_string_getter).unwrap();
+
+        let mut test_path = config.diary_path().clone();
+        test_path.push("2021-11/diary_2021-11-06.md");
+
+        let content = fs::read_to_string(test_path).unwrap();
+        assert!(content.contains("Test content"));
     }
 }