+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+### Changed
+
+* New command errors if diary is unintialised.
+
 ## [0.2.0] - 2021-11-10
 
 ### Added